+++ misskey-websocket/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+- Use tokio 1.0 and async-tungstenite 0.13
+  - Drop feature flags for tokio 0.2 and async-tungstenite 0.9
+
 ### Deprecated
 ### Removed
 ### Fixed